# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Biglands API
  description: |
    Biglands — Bất động sản CHDV platform. API for managing property listings,
    deposits, deal lifecycle, user management, approvals, and notifications.

    **Base URL**: `/api/v1`
    **Auth**: Bearer JWT (24h expiry)
    **Pagination**: offset-based (`page`, `size`)
  version: "1.0.0"
servers:
  - url: /api/v1
    description: Biglands API v1
paths:
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List current user's notifications (newest first)
      operationId: listNotifications
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - name: isRead
          in: query
          schema:
            type: boolean
          description: Filter by read status
          required: false
      responses:
        "200":
          description: Paginated notification list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/unread-count:
    get:
      tags: [Notifications]
      summary: Get unread notification count
      operationId: getUnreadCount
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    minimum: 0
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Mark a single notification as read
      operationId: markNotificationRead
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/notificationId"
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /notifications/read-all:
    post:
      tags: [Notifications]
      summary: Mark all current user's notifications as read
      operationId: markAllNotificationsRead
      security:
        - BearerAuth: []
      responses:
        "200":
          description: All notifications marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
                    description: Number of notifications updated
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ──────────────────────────────────────────────
  #  USER PINS
  # ──────────────────────────────────────────────
  /listings/{id}/pin:
    put:
      tags: [Pins]
      summary: Pin a listing (toggle; idempotent)
      operationId: pinListing
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
      responses:
        "200":
          description: Listing pinned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPin"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Pins]
      summary: Unpin a listing (idempotent)
      operationId: unpinListing
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
      responses:
        "204":
          description: Listing unpinned
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/me/pins:
    get:
      tags: [Pins]
      summary: List current user's pinned listings
      operationId: listMyPins
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
      responses:
        "200":
          description: Paginated list of pinned listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ──────────────────────────────────────────────
  #  USER SETTINGS (ME)
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer token. Issued via POST /auth/login. Expires in 24h."
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
      required: false
    size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
      required: false
...
    listingId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Listing UUID
...
    notificationId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Notification UUID
  responses:
...
    Unauthorized:
      description: "Missing, invalid, or expired authentication token"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    NotFound:
      description: "Requested resource was not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    TransactionType:
      type: string
      enum: [SANG_NHUONG, CHO_THUE, BAN]
      description: "Transaction type for a listing."
    ListingStatus:
      type: string
      enum: [DRAFT, PENDING_APPROVAL, CON_HANG, DA_COC, HET_HANG, DA_CHOT, HUY_COC, QUA_HAN]
      description: "Listing status machine. See domain-model.md §11 for state transitions."
    CommissionType:
      type: string
      enum: [PERCENTAGE, FLAT]
      description: "Required per BR-008, SC-004. PERCENTAGE = % of price; FLAT = fixed VND amount."
    PropertyType:
      type: string
      enum: [NHA_PHO, CAN_HO, CHDV, DAT, BIET_THU, VAN_PHONG, MAT_BANG, KHO_XUONG, NHA_TRO, KHAC]
      description: "Property type categories from create form Loại field."
...
    ReferenceType:
      type: string
      enum: [LISTING, APPROVAL, DEAL_EVENT]
...
    Listing:
      type: object
      required:
        - id
        - code
        - transactionType
        - propertyType
        - description
        - price
        - commissionType
        - commissionValue
        - areaWidth
        - areaLength
        - totalArea
        - numRooms
        - numBathrooms
        - numFloors
        - streetName
        - houseNumber
        - address
        - ward
        - district
        - city
        - ownerPhone
        - status
        - createdById
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 20
          description: "Auto-generated: YYMMDD + random digits (LST-C01)"
        transactionType:
          $ref: "#/components/schemas/TransactionType"
        propertyType:
          $ref: "#/components/schemas/PropertyType"
        title:
          type: string
          maxLength: 500
          nullable: true
          description: "Optional title with optional prefix (NNC, MT, HXH, CHDV)"
        description:
          type: string
          description: Property description
        price:
          type: number
          description: Price in VND
        commissionType:
          $ref: "#/components/schemas/CommissionType"
        commissionValue:
          type: number
          description: "Percentage (e.g., 1.5) or fixed VND amount"
        areaWidth:
          type: number
          format: float
          description: Frontage width in meters
        areaLength:
          type: number
          format: float
          description: Property length in meters
        totalArea:
          type: number
          format: float
          description: "Total floor area in m² (may be computed)"
        numRooms:
          type: integer
          minimum: 0
          description: Number of bedrooms
        numBathrooms:
          type: integer
          minimum: 0
          description: Number of bathrooms/toilets
        numFloors:
          type: integer
          minimum: 0
          description: Number of floors
        streetName:
          type: string
          maxLength: 255
        houseNumber:
          type: string
          maxLength: 50
        address:
          type: string
          maxLength: 500
          description: "Full composite address"
        ward:
          type: string
          maxLength: 100
        district:
          type: string
          maxLength: 100
        city:
          type: string
          maxLength: 100
          default: "Hồ Chí Minh"
        latitude:
          type: number
          format: float
          nullable: true
        longitude:
          type: number
          format: float
          nullable: true
        label:
          type: string
          maxLength: 100
          nullable: true
          description: "e.g., Thang máy, Nhà mới, Vị trí đẹp"
        furnishing:
          type: string
          maxLength: 500
          nullable: true
          description: Furniture description
        frontageType:
          type: string
          maxLength: 100
          nullable: true
          description: "Mặt tiền / Hẻm"
        legalStatus:
          type: string
          maxLength: 500
          nullable: true
          description: Legal documentation
        direction:
          type: string
          maxLength: 50
          nullable: true
          description: "Hướng — e.g., Đông, Tây, Nam, Bắc"
        roadWidth:
          type: string
          maxLength: 50
          nullable: true
          description: Alley/road width
        ownerPhone:
          type: string
          maxLength: 20
          description: Property owner's contact phone
        videoUrl:
          type: string
          format: uri
          maxLength: 500
          nullable: true
          description: "YouTube video link (max 1 per listing — BR-014)"
        status:
          $ref: "#/components/schemas/ListingStatus"
        isHot:
          type: boolean
          default: false
          description: Promoted to Hot by Admin
        hotOrder:
          type: integer
          nullable: true
          description: Display order in Hot section
        viewCount:
          type: integer
          default: 0
        createdById:
          type: string
          format: uuid
          description: Agent who created the listing
        approvedById:
          type: string
          format: uuid
          nullable: true
          description: Approver who approved
        approvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ListingListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        pagination:
          $ref: "#/components/schemas/Pagination"
        totalCount:
          type: integer
          description: "Total count of CON_HANG listings across all transaction types (BR-015)"

    # ── ListingImage ───────────────────────────
...
    Notification:
      type: object
      required: [id, userId, title, body, isRead, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 500
        body:
          type: string
        referenceType:
          $ref: "#/components/schemas/ReferenceType"
          nullable: true
          description: "Polymorphic reference (NTF-C01)"
        referenceId:
          type: string
          format: uuid
          nullable: true
          description: "References listing, approval, or deal event (NTF-C01)"
        isRead:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
    NotificationListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    UserPin:
      type: object
      required: [userId, listingId, pinnedAt]
      properties:
        userId:
          type: string
          format: uuid
        listingId:
          type: string
          format: uuid
        pinnedAt:
          type: string
          format: date-time

    # ── Hot Products ──────────────────────────
...
    Pagination:
      type: object
      required: [page, size, totalItems, totalPages]
      properties:
        page:
          type: integer
          description: Current page number
        size:
          type: integer
          description: Items per page
        totalItems:
          type: integer
          description: Total number of items matching the query
        totalPages:
          type: integer
          description: Total number of pages
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: "Error code — e.g., VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, ALREADY_PROCESSED, INVALID_STATUS_TRANSITION"
        message:
          type: string
          description: Human-readable error description
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: "Field path for validation errors"
              message:
                type: string
                description: "Field-level error message"
              code:
                type: string
                description: "Field-level error code: required, minLength, maxLength, pattern, minimum, maximum, invalid, duplicate, conflict"